database_gateway/requirements.txt
motor==3.3.*
grpcio==1.59.*
protobuf==4.24.*
